Rename GitForm to GitSettingsForm update GitConfigurationDialog
